libnozzle: refactor platform-specific code into dedicated files
Split platform-specific implementation from libnozzle.c into dedicated
files: libnozzle_linux.c, libnozzle_bsd.c, libnozzle_solaris.c.

Introduce platform abstraction layer with typedefs and macros in internals.h:
- nozzle_ifreq: Platform-specific ifreq structure type
- NOZZLE_SOCKET_DOMAIN: Socket domain for ioctl fd (AF_LOCAL on BSD)
- NOZZLE_IOCTL_FD: File descriptor for ioctl operations
- NOZZLE_SET_MTU: Platform-specific MTU ioctl command
- NOZZLE_IPV6_IS_SECONDARY(domain): IPv6 secondary interface check

This eliminates all platform-specific #ifdef blocks from libnozzle.c,
improving code organization and readability. Platform differences are
now documented in a single location (internals.h and platform-specific
files) rather than scattered throughout the main code.
