+++ libnozzle/internals.h
@@ -11,6 +11,8 @@
 
 #include "config.h"
 
+#include <limits.h>
+
 #ifdef KNET_LINUX
 #include <netlink/netlink.h>
 #endif
@@ -26,7 +28,7 @@ struct nozzle_lib_config {
 #ifdef KNET_LINUX
 	struct nl_sock *nlsock;
 #endif
-#ifdef KNET_BSD
+#if defined(KNET_BSD) || defined(KNET_SOLARIS)
 	int ip_fd;
 	int ip6_fd;
 #endif
@@ -63,20 +65,44 @@ struct nozzle_iface {
 	struct nozzle_iface *next;
 
 #ifdef KNET_SOLARIS
-	int ip_fd;
-	int ip6_fd;
+	int ip_fd;	/* STREAMS plumbing socket for IPv4 */
+	int ip6_fd;	/* STREAMS plumbing socket for IPv6 */
 #endif
 };
 
+/*
+ * Platform-specific type and macro abstractions
+ */
 #ifdef KNET_SOLARIS
+typedef struct lifreq nozzle_ifreq;
 #define ifname ifr.lifr_name
 #define ifmtu ifr.lifr_mtu
 #define ifflags ifr.lifr_flags
+#define NOZZLE_IOCTL_FD lib_cfg.ip_fd
+#define NOZZLE_SET_MTU SIOCSLIFMTU
+#define NOZZLE_SOCKET_DOMAIN AF_INET
+#define NOZZLE_IPV6_IS_SECONDARY(domain) ((domain) == AF_INET6)
 #else
+typedef struct ifreq nozzle_ifreq;
 #define ifname ifr.ifr_name
 #define ifmtu ifr.ifr_mtu
 #define ifflags ifr.ifr_flags
+#define NOZZLE_SET_MTU SIOCSIFMTU
+#define NOZZLE_SOCKET_DOMAIN AF_INET
+#define NOZZLE_IPV6_IS_SECONDARY(domain) 0
+#ifdef KNET_BSD
+#undef NOZZLE_SOCKET_DOMAIN
+#define NOZZLE_SOCKET_DOMAIN AF_LOCAL
+#define NOZZLE_IOCTL_FD lib_cfg.ioctlfd
+#else
+#define NOZZLE_IOCTL_FD lib_cfg.ioctlfd
 #endif
+#endif
+
+/*
+ * Global library configuration (defined in libnozzle.c)
+ */
+extern struct nozzle_lib_config lib_cfg;
 
 int execute_bin_sh_command(const char *command, char **error_string);
 
@@ -85,5 +111,22 @@ int find_ip(nozzle_t nozzle,
 	    struct nozzle_ip **ip, struct nozzle_ip **ip_prev);
 
 char *generate_v4_broadcast(const char *ipaddr, const char *prefix);
+
+/*
+ * Platform-specific functions
+ */
+int _platform_init(struct nozzle_lib_config *lib_cfg);
+void _platform_fini(struct nozzle_lib_config *lib_cfg);
+
+int _platform_create_tap(nozzle_t nozzle, char *devname, size_t devname_size);
+void _platform_close_tap(nozzle_t nozzle);
+void _platform_destroy_tap(nozzle_t nozzle);
+
+int _platform_get_mac(const nozzle_t nozzle, char **ether_addr);
+int _platform_set_mac(nozzle_t nozzle, const char *ether_addr);
+int _platform_get_mtu(const nozzle_t nozzle);
+
+int _platform_add_ip(nozzle_t nozzle, const char *ipaddr, const char *prefix, int secondary);
+int _platform_del_ip(nozzle_t nozzle, const char *ipaddr, const char *prefix, int secondary);
 
 #endif
