+++ libnozzle/libnozzle.c
@@ -13,54 +13,16 @@
 #include <string.h>
 #include <sys/types.h>
 #include <sys/stat.h>
-#include <sys/socket.h>
 #include <sys/wait.h>
 #include <fcntl.h>
 #include <unistd.h>
 #include <sys/ioctl.h>
-#ifdef KNET_SOLARIS
-#include <sys/sockio.h>
-#include <sys/ethernet.h>
-#include <net/if_types.h>
-#include <net/if_tun.h>
-#include <stropts.h>
-#include <libdlpi.h>
-#include <netinet/in.h>
-#include <arpa/inet.h>
-#else
-#include <net/ethernet.h>
-#endif
+#include <sys/socket.h>
 #include <pthread.h>
 #include <limits.h>
 #include <stdio.h>
 #include <net/if.h>
-#include <ifaddrs.h>
 #include <stdint.h>
-
-#ifdef KNET_LINUX
-#include <linux/if_tun.h>
-/*
- * libnl3 < 3.3 includes kernel headers directly
- * causing conflicts with net/if.h included above
- */
-#ifdef LIBNL3_WORKAROUND
-#define _LINUX_IF_H 1
-#endif
-#include <netinet/ether.h>
-#include <netlink/netlink.h>
-#include <netlink/route/addr.h>
-#include <netlink/route/link.h>
-#include <netlink/errno.h>
-#endif
-#ifdef KNET_BSD
-#include <net/if_dl.h>
-#include <sys/sockio.h>
-#include <netinet/in.h>
-#include <netinet/in_var.h>
-#include <netinet6/in6_var.h>
-#include <netinet6/nd6.h>
-#include <arpa/inet.h>
-#endif
 
 #include "libnozzle.h"
 #include "internals.h"
@@ -70,73 +32,18 @@
  * locking should be handled at external API functions
  */
 static int lib_init = 0;
-static struct nozzle_lib_config lib_cfg;
+struct nozzle_lib_config lib_cfg;
 static pthread_mutex_t config_mutex = PTHREAD_MUTEX_INITIALIZER;
 
 /*
  * internal helpers
  */
 
-#ifdef KNET_LINUX
-/*
- * Convert libnl error codes to errno values for better error reporting.
- * libnl functions return negative NLE_* error codes, which need translation
- * to standard errno values that applications expect.
- */
-static int nlerr_to_errno(int nlerr)
-{
-	if (nlerr >= 0)
-		return 0;
-
-	/*
-	 * NLE_* error codes are small negative integers.
-	 * Kernel errors passed through netlink are already errno values.
-	 * Use NLE_MAX as threshold to distinguish between the two.
-	 */
-	if (-nlerr > NLE_MAX) {
-		/* Already an errno value from kernel */
-		return -nlerr;
-	}
-
-	/* Map common NLE_* codes to errno values */
-	switch (-nlerr) {
-	case NLE_NOMEM:
-		return ENOMEM;
-	case NLE_EXIST:
-		return EEXIST;
-	case NLE_INVAL:
-		return EINVAL;
-	case NLE_NOADDR:
-		return EADDRNOTAVAIL;
-	case NLE_NODEV:
-		return ENODEV;
-	case NLE_BUSY:
-		return EBUSY;
-	case NLE_PERM:
-		return EPERM;
-	case NLE_OBJ_NOTFOUND:
-		return ENOENT;
-	case NLE_RANGE:
-		return ERANGE;
-	default:
-		/* Generic fallback for unmapped errors */
-		return EINVAL;
-	}
-}
-#endif
-
 static void lib_fini(void)
 {
 	if (lib_cfg.head == NULL) {
-#ifdef KNET_LINUX
-		nl_close(lib_cfg.nlsock);
-		nl_socket_free(lib_cfg.nlsock);
-#endif
+		_platform_fini(&lib_cfg);
 		close(lib_cfg.ioctlfd);
-#ifdef KNET_BSD
-		close(lib_cfg.ip_fd);
-		close(lib_cfg.ip6_fd);
-#endif
 		lib_init = 0;
 	}
 }
@@ -167,23 +74,13 @@ static int is_valid_nozzle(const nozzle_t nozzle)
 
 static void destroy_iface(nozzle_t nozzle)
 {
-#ifdef KNET_BSD
-	struct ifreq ifr;
-#endif
-
 	if (!nozzle)
 		return;
 
 	if (nozzle->fd >= 0)
 		close(nozzle->fd);
 
-#ifdef KNET_BSD
-	memset(&ifr, 0, sizeof(struct ifreq));
-	memmove(ifname, nozzle->name, IFNAMSIZ);
-
-	ioctl(lib_cfg.ioctlfd, SIOCIFDESTROY, &ifr);
-	ioctl(lib_cfg.ioctlfd, SIOCGIFFLAGS, &ifr);
-#endif
+	_platform_destroy_tap(nozzle);
 
 	free(nozzle);
 
@@ -194,122 +91,12 @@ static void destroy_iface(nozzle_t nozzle)
 
 static int get_iface_mtu(const nozzle_t nozzle)
 {
-	int err = 0, savederrno = 0;
-#ifdef KNET_SOLARIS
-	struct lifreq ifr;
-#else
-	struct ifreq ifr;
-#endif
-	memset(&ifr, 0, sizeof(ifr));
-	memmove(ifname, nozzle->name, IFNAMSIZ);
-
-#ifdef KNET_SOLARIS
-	err = ioctl(lib_cfg.ioctlfd, SIOCGLIFMTU, &ifr);
-#else
-	err = ioctl(lib_cfg.ioctlfd, SIOCGIFMTU, &ifr);
-#endif
-	if (err) {
-		savederrno = errno;
-		goto out_clean;
-	}
-
-	err = ifmtu;
-out_clean:
-	errno = savederrno;
-	return err;
+	return _platform_get_mtu(nozzle);
 }
 
 static int get_iface_mac(const nozzle_t nozzle, char **ether_addr)
 {
-	int err = 0, savederrno = 0;
-#ifdef  KNET_SOLARIS
-	struct lifreq ifr;
-	dlpi_handle_t dlpi_handle;
-	dlpi_info_t dlpi_if_info;
-#else
-	struct ifreq ifr;
-#endif
-	char mac[MACADDR_CHAR_MAX];
-#ifdef KNET_BSD
-	struct ifaddrs *ifap = NULL;
-	struct ifaddrs *ifa;
-	int found = 0;
-#endif
-
-	memset(&mac, 0, MACADDR_CHAR_MAX);
-	memset(&ifr, 0, sizeof(struct ifreq));
-	memmove(ifname, nozzle->name, IFNAMSIZ);
-
-#ifdef KNET_LINUX
-	err = ioctl(lib_cfg.ioctlfd, SIOCGIFHWADDR, &ifr);
-	if (err) {
-		savederrno = errno;
-		goto out_clean;
-	}
-
-	ether_ntoa_r((struct ether_addr *)ifr.ifr_hwaddr.sa_data, mac);
-#endif
-#ifdef KNET_SOLARIS
-	err = dlpi_open(nozzle->name, &dlpi_handle, 0);
-	if (err != DLPI_SUCCESS) {
-		err = -1;
-		goto out_clean;
-
-	}
-	err = dlpi_info(dlpi_handle, &dlpi_if_info, 0);
-
-	if (err != DLPI_SUCCESS) {
-		err = -1;
-		dlpi_close(dlpi_handle);
-		goto out_clean;
-	}
-	dlpi_close(dlpi_handle);
-	ether_ntoa_r((struct ether_addr *)dlpi_if_info.di_physaddr, mac);
-#endif
-#ifdef KNET_BSD
-	/*
-	 * there is no ioctl to get the ether address of an interface on FreeBSD
-	 * (not to be confused with hwaddr). Use workaround described here:
-	 * https://lists.freebsd.org/pipermail/freebsd-hackers/2004-June/007394.html
-	 */
-	err = getifaddrs(&ifap);
-	if (err < 0) {
-		savederrno = errno;
-		goto out_clean;
-	}
-
-	ifa = ifap;
-
-	while (ifa) {
-		if (!strncmp(nozzle->name, ifa->ifa_name, IFNAMSIZ)) {
-			found = 1;
-			break;
-		}
-		ifa=ifa->ifa_next;
-	}
-
-	if (found) {
-		ether_ntoa_r((struct ether_addr *)LLADDR((struct sockaddr_dl *)ifa->ifa_addr), mac);
-	} else {
-		errno = EINVAL;
-		err = -1;
-	}
-
-	freeifaddrs(ifap);
-
-	if (err) {
-		goto out_clean;
-	}
-
-#endif
-	*ether_addr = strdup(mac);
-	if (!*ether_addr) {
-		savederrno = errno;
-		err = -1;
-	}
-out_clean:
-	errno = savederrno;
-	return err;
+	return _platform_get_mac(nozzle, ether_addr);
 }
 
 #define IP_ADD 1
@@ -320,597 +107,12 @@ static int _set_ip(nozzle_t nozzle,
 		   const char *ipaddr, const char *prefix,
 		   int secondary)
 {
-	int fam;
-	int err = 0;
-	char *broadcast = NULL;
-#ifdef KNET_LINUX
-	struct rtnl_addr *addr = NULL;
-	struct nl_addr *local_addr = NULL;
-	struct nl_addr *bcast_addr = NULL;
-	struct nl_cache *cache = NULL;
-	int ifindex;
-#endif
-
-	if (!strchr(ipaddr, ':')) {
-		fam = AF_INET;
-		broadcast = generate_v4_broadcast(ipaddr, prefix);
-		if (!broadcast) {
-			errno = EINVAL;
-			return -1;
-		}
-	} else {
-		fam = AF_INET6;
-	}
-
-#ifdef KNET_LINUX
-	int nlerr;
-
-	addr = rtnl_addr_alloc();
-	if (!addr) {
-		errno = ENOMEM;
-		err = -1;
-		goto out;
-	}
-
-	nlerr = rtnl_link_alloc_cache(lib_cfg.nlsock, AF_UNSPEC, &cache);
-	if (nlerr < 0) {
-		errno = nlerr_to_errno(nlerr);
-		err = -1;
-		goto out;
-	}
-
-	ifindex = rtnl_link_name2i(cache, nozzle->name);
-	if (ifindex == 0) {
-		errno = ENOENT;
-		err = -1;
-		goto out;
-	}
-
-	rtnl_addr_set_ifindex(addr, ifindex);
-
-	nlerr = nl_addr_parse(ipaddr, fam, &local_addr);
-	if (nlerr < 0) {
-		errno = nlerr_to_errno(nlerr);
-		err = -1;
-		goto out;
-	}
-
-	nlerr = rtnl_addr_set_local(addr, local_addr);
-	if (nlerr < 0) {
-		errno = nlerr_to_errno(nlerr);
-		err = -1;
-		goto out;
-	}
-
-	if (broadcast) {
-		nlerr = nl_addr_parse(broadcast, fam, &bcast_addr);
-		if (nlerr < 0) {
-			errno = nlerr_to_errno(nlerr);
-			err = -1;
-			goto out;
-		}
-
-		nlerr = rtnl_addr_set_broadcast(addr, bcast_addr);
-		if (nlerr < 0) {
-			errno = nlerr_to_errno(nlerr);
-			err = -1;
-			goto out;
-		}
-	}
-
-	rtnl_addr_set_prefixlen(addr, atoi(prefix));
-
 	if (command == IP_ADD) {
-		nlerr = rtnl_addr_add(lib_cfg.nlsock, addr, 0);
-		if (nlerr < 0) {
-			errno = nlerr_to_errno(nlerr);
-			err = -1;
-			goto out;
-		}
+		return _platform_add_ip(nozzle, ipaddr, prefix, secondary);
 	} else {
-		nlerr = rtnl_addr_delete(lib_cfg.nlsock, addr, 0);
-		if (nlerr < 0) {
-			errno = nlerr_to_errno(nlerr);
-			err = -1;
-			goto out;
-		}
+		return _platform_del_ip(nozzle, ipaddr, prefix, secondary);
 	}
-out:
-	if (addr) {
-		rtnl_addr_put(addr);
-	}
-	if (local_addr) {
-		nl_addr_put(local_addr);
-	}
-	if (bcast_addr) {
-		nl_addr_put(bcast_addr);
-	}
-	if (cache) {
-		nl_cache_put(cache);
-	}
-	if (broadcast) {
-		free(broadcast);
-	}
-	return err;
-#endif
-#ifdef KNET_BSD
-	/*
-	 * FreeBSD implementation using ioctl
-	 */
-	if (fam == AF_INET) {
-		struct in_aliasreq ifra;
-		int prefix_len;
-		uint32_t mask;
-
-		memset(&ifra, 0, sizeof(ifra));
-		strncpy(ifra.ifra_name, nozzle->name, IFNAMSIZ);
-
-		/* Parse prefix length */
-		prefix_len = atoi(prefix);
-		if (prefix_len <= 0 || prefix_len > 32) {
-			if (broadcast) {
-				free(broadcast);
-			}
-			errno = EINVAL;
-			return -1;
-		}
-
-		if (command == IP_ADD) {
-			/* Set address */
-			ifra.ifra_addr.sin_family = AF_INET;
-			ifra.ifra_addr.sin_len = sizeof(struct sockaddr_in);
-			if (inet_pton(AF_INET, ipaddr, &ifra.ifra_addr.sin_addr) <= 0) {
-				if (broadcast) {
-					free(broadcast);
-				}
-				errno = EINVAL;
-				return -1;
-			}
-
-			/* Set netmask */
-			ifra.ifra_mask.sin_family = AF_INET;
-			ifra.ifra_mask.sin_len = sizeof(struct sockaddr_in);
-			mask = htonl(~((1 << (32 - prefix_len)) - 1));
-			ifra.ifra_mask.sin_addr.s_addr = mask;
-
-			/* Set broadcast address */
-			if (broadcast) {
-				ifra.ifra_broadaddr.sin_family = AF_INET;
-				ifra.ifra_broadaddr.sin_len = sizeof(struct sockaddr_in);
-				if (inet_pton(AF_INET, broadcast, &ifra.ifra_broadaddr.sin_addr) <= 0) {
-					free(broadcast);
-					errno = EINVAL;
-					return -1;
-				}
-			}
-
-			err = ioctl(lib_cfg.ip_fd, SIOCAIFADDR, &ifra);
-		} else {
-			/* Delete address */
-			struct ifreq ifr;
-			struct sockaddr_in *sin;
-
-			memset(&ifr, 0, sizeof(ifr));
-			strncpy(ifr.ifr_name, nozzle->name, IFNAMSIZ);
-
-			sin = (struct sockaddr_in *)&ifr.ifr_addr;
-			sin->sin_family = AF_INET;
-			sin->sin_len = sizeof(struct sockaddr_in);
-			if (inet_pton(AF_INET, ipaddr, &sin->sin_addr) <= 0) {
-				if (broadcast) {
-					free(broadcast);
-				}
-				errno = EINVAL;
-				return -1;
-			}
-
-			err = ioctl(lib_cfg.ip_fd, SIOCDIFADDR, &ifr);
-		}
-
-		if (broadcast) {
-			free(broadcast);
-		}
-	} else {
-		/* IPv6 */
-		struct in6_aliasreq ifra6;
-		struct sockaddr_in6 *sin6_addr, *sin6_mask;
-		int prefix_len;
-		int i;
-
-		memset(&ifra6, 0, sizeof(ifra6));
-		strncpy(ifra6.ifra_name, nozzle->name, IFNAMSIZ);
-
-		/* Parse prefix length */
-		prefix_len = atoi(prefix);
-		if (prefix_len <= 0 || prefix_len > 128) {
-			errno = EINVAL;
-			return -1;
-		}
-
-		if (command == IP_ADD) {
-			/* Set address */
-			sin6_addr = &ifra6.ifra_addr;
-			sin6_addr->sin6_family = AF_INET6;
-			sin6_addr->sin6_len = sizeof(struct sockaddr_in6);
-			if (inet_pton(AF_INET6, ipaddr, &sin6_addr->sin6_addr) <= 0) {
-				errno = EINVAL;
-				return -1;
-			}
-
-			/* Set prefix mask */
-			sin6_mask = &ifra6.ifra_prefixmask;
-			sin6_mask->sin6_family = AF_INET6;
-			sin6_mask->sin6_len = sizeof(struct sockaddr_in6);
-
-			/* Convert prefix length to mask */
-			for (i = 0; i < 16; i++) {
-				if (prefix_len >= 8) {
-					sin6_mask->sin6_addr.s6_addr[i] = 0xff;
-					prefix_len -= 8;
-				} else if (prefix_len > 0) {
-					sin6_mask->sin6_addr.s6_addr[i] = (0xff << (8 - prefix_len)) & 0xff;
-					prefix_len = 0;
-				} else {
-					sin6_mask->sin6_addr.s6_addr[i] = 0;
-				}
-			}
-
-			/* Set address lifetime to infinity */
-			ifra6.ifra_lifetime.ia6t_vltime = ND6_INFINITE_LIFETIME;
-			ifra6.ifra_lifetime.ia6t_pltime = ND6_INFINITE_LIFETIME;
-
-			err = ioctl(lib_cfg.ip6_fd, SIOCAIFADDR_IN6, &ifra6);
-		} else {
-			/* Delete address */
-			struct in6_ifreq ifr6;
-
-			memset(&ifr6, 0, sizeof(ifr6));
-			strncpy(ifr6.ifr_name, nozzle->name, IFNAMSIZ);
-
-			ifr6.ifr_addr.sin6_family = AF_INET6;
-			ifr6.ifr_addr.sin6_len = sizeof(struct sockaddr_in6);
-			if (inet_pton(AF_INET6, ipaddr, &ifr6.ifr_addr.sin6_addr) <= 0) {
-				errno = EINVAL;
-				return -1;
-			}
-
-			err = ioctl(lib_cfg.ip6_fd, SIOCDIFADDR_IN6, &ifr6);
-		}
-	}
-
-	if (err < 0) {
-		err = -1;
-	}
-
-	return err;
-#endif
-#ifdef KNET_SOLARIS
-	/*
-	 * Solaris implementation using ioctl
-	 */
-	if (fam == AF_INET) {
-		struct lifreq lifr;
-		struct sockaddr_in *sin_addr, *sin_mask, *sin_bcast;
-		int prefix_len;
-		uint32_t mask;
-
-		memset(&lifr, 0, sizeof(lifr));
-		strncpy(lifr.lifr_name, nozzle->name, LIFNAMSIZ);
-
-		/* Parse prefix length */
-		prefix_len = atoi(prefix);
-		if (prefix_len <= 0 || prefix_len > 32) {
-			if (broadcast) {
-				free(broadcast);
-			}
-			errno = EINVAL;
-			return -1;
-		}
-
-		if (command == IP_ADD) {
-			/* For secondary addresses, create logical interface first */
-			if (secondary) {
-				/* SIOCLIFADDIF with zero address */
-				memset(&lifr.lifr_addr, 0, sizeof(lifr.lifr_addr));
-				if (ioctl(nozzle->ip_fd, SIOCLIFADDIF, &lifr) < 0) {
-					if (broadcast) {
-						free(broadcast);
-					}
-					return -1;
-				}
-				/* lifr_name is updated by SIOCLIFADDIF with new logical if name */
-			}
-
-			/* Set netmask */
-			sin_mask = (struct sockaddr_in *)&lifr.lifr_addr;
-			sin_mask->sin_family = AF_INET;
-			mask = htonl(~((1 << (32 - prefix_len)) - 1));
-			sin_mask->sin_addr.s_addr = mask;
-
-			if (ioctl(nozzle->ip_fd, SIOCSLIFNETMASK, &lifr) < 0) {
-				err = -1;
-				if (secondary) {
-					/* Clean up the logical interface we created */
-					(void)ioctl(nozzle->ip_fd, SIOCLIFREMOVEIF, &lifr);
-				}
-				if (broadcast) {
-					free(broadcast);
-				}
-				return err;
-			}
-
-			/* Set address */
-			sin_addr = (struct sockaddr_in *)&lifr.lifr_addr;
-			sin_addr->sin_family = AF_INET;
-			if (inet_pton(AF_INET, ipaddr, &sin_addr->sin_addr) <= 0) {
-				if (secondary) {
-					(void)ioctl(nozzle->ip_fd, SIOCLIFREMOVEIF, &lifr);
-				}
-				if (broadcast) {
-					free(broadcast);
-				}
-				errno = EINVAL;
-				return -1;
-			}
-
-			err = ioctl(nozzle->ip_fd, SIOCSLIFADDR, &lifr);
-			if (err < 0) {
-				if (secondary) {
-					(void)ioctl(nozzle->ip_fd, SIOCLIFREMOVEIF, &lifr);
-				}
-				if (broadcast) {
-					free(broadcast);
-				}
-				return -1;
-			}
-
-			/* Set broadcast address if provided */
-			if (broadcast) {
-				sin_bcast = (struct sockaddr_in *)&lifr.lifr_broadaddr;
-				sin_bcast->sin_family = AF_INET;
-				if (inet_pton(AF_INET, broadcast, &sin_bcast->sin_addr) <= 0) {
-					free(broadcast);
-					errno = EINVAL;
-					return -1;
-				}
-
-				err = ioctl(nozzle->ip_fd, SIOCSLIFBRDADDR, &lifr);
-				free(broadcast);
-				/* Non-fatal if broadcast fails */
-			}
-		} else {
-			/* Delete address */
-			if (secondary) {
-				/* Set the address to delete so Solaris can find the right logical interface */
-				sin_addr = (struct sockaddr_in *)&lifr.lifr_addr;
-				sin_addr->sin_family = AF_INET;
-				if (inet_pton(AF_INET, ipaddr, &sin_addr->sin_addr) <= 0) {
-					if (broadcast) {
-						free(broadcast);
-					}
-					errno = EINVAL;
-					return -1;
-				}
-				err = ioctl(nozzle->ip_fd, SIOCLIFREMOVEIF, &lifr);
-			} else {
-				/* Primary IPv4: set to 0.0.0.0 */
-				sin_addr = (struct sockaddr_in *)&lifr.lifr_addr;
-				sin_addr->sin_family = AF_INET;
-				sin_addr->sin_addr.s_addr = 0;
-				err = ioctl(nozzle->ip_fd, SIOCSLIFADDR, &lifr);
-			}
-
-			if (broadcast) {
-				free(broadcast);
-			}
-		}
-	} else {
-		/* IPv6 */
-		struct lifreq lifr;
-		struct sockaddr_in6 *sin6_addr, *sin6_mask;
-		int prefix_len;
-		int i;
-
-		memset(&lifr, 0, sizeof(lifr));
-		strncpy(lifr.lifr_name, nozzle->name, LIFNAMSIZ);
-
-		/* Parse prefix length */
-		prefix_len = atoi(prefix);
-		if (prefix_len <= 0 || prefix_len > 128) {
-			errno = EINVAL;
-			return -1;
-		}
-
-		if (command == IP_ADD) {
-			/* For secondary addresses, create logical interface first */
-			if (secondary) {
-				/* SIOCLIFADDIF with zero address */
-				memset(&lifr.lifr_addr, 0, sizeof(lifr.lifr_addr));
-				if (ioctl(nozzle->ip6_fd, SIOCLIFADDIF, &lifr) < 0) {
-					return -1;
-				}
-			}
-
-			/* Set prefix mask */
-			sin6_mask = (struct sockaddr_in6 *)&lifr.lifr_addr;
-			sin6_mask->sin6_family = AF_INET6;
-
-			/* Convert prefix length to mask */
-			for (i = 0; i < 16; i++) {
-				if (prefix_len >= 8) {
-					sin6_mask->sin6_addr.s6_addr[i] = 0xff;
-					prefix_len -= 8;
-				} else if (prefix_len > 0) {
-					sin6_mask->sin6_addr.s6_addr[i] = (0xff << (8 - prefix_len)) & 0xff;
-					prefix_len = 0;
-				} else {
-					sin6_mask->sin6_addr.s6_addr[i] = 0;
-				}
-			}
-
-			if (ioctl(nozzle->ip6_fd, SIOCSLIFNETMASK, &lifr) < 0) {
-				if (secondary) {
-					(void)ioctl(nozzle->ip6_fd, SIOCLIFREMOVEIF, &lifr);
-				}
-				return -1;
-			}
-
-			/* Reset prefix_len for address setting */
-			prefix_len = atoi(prefix);
-
-			/* Set address */
-			sin6_addr = (struct sockaddr_in6 *)&lifr.lifr_addr;
-			sin6_addr->sin6_family = AF_INET6;
-			if (inet_pton(AF_INET6, ipaddr, &sin6_addr->sin6_addr) <= 0) {
-				if (secondary) {
-					(void)ioctl(nozzle->ip6_fd, SIOCLIFREMOVEIF, &lifr);
-				}
-				errno = EINVAL;
-				return -1;
-			}
-
-			err = ioctl(nozzle->ip6_fd, SIOCSLIFADDR, &lifr);
-			if (err < 0) {
-				if (secondary) {
-					(void)ioctl(nozzle->ip6_fd, SIOCLIFREMOVEIF, &lifr);
-				}
-				return -1;
-			}
-		} else {
-			/* Delete address */
-			if (secondary) {
-				/* Set the address to delete so Solaris can find the right logical interface */
-				sin6_addr = (struct sockaddr_in6 *)&lifr.lifr_addr;
-				sin6_addr->sin6_family = AF_INET6;
-				if (inet_pton(AF_INET6, ipaddr, &sin6_addr->sin6_addr) <= 0) {
-					errno = EINVAL;
-					return -1;
-				}
-				err = ioctl(nozzle->ip6_fd, SIOCLIFREMOVEIF, &lifr);
-			} else {
-				/* Primary IPv6: set to :: */
-				sin6_addr = (struct sockaddr_in6 *)&lifr.lifr_addr;
-				sin6_addr->sin6_family = AF_INET6;
-				memset(&sin6_addr->sin6_addr, 0, sizeof(sin6_addr->sin6_addr));
-				err = ioctl(nozzle->ip6_fd, SIOCSLIFADDR, &lifr);
-			}
-		}
-	}
-
-	if (err < 0) {
-		err = -1;
-	}
-
-	return err;
-#endif
 }
-
-
-#ifdef KNET_SOLARIS
-// Most of this taken from openconnect
-static int link_proto(nozzle_t nozzle, int unit_nr,
-		      const char *devname, uint64_t flags)
-{
-	int ip_fd, mux_id, tap2_fd;
-	struct lifreq ifr;
-
-	tap2_fd = open("/dev/tap", O_RDWR);
-	if (tap2_fd < 0) {
-		return -EIO;
-	}
-	if (ioctl(tap2_fd, I_PUSH, "ip") < 0) {
-		close(tap2_fd);
-		return -EIO;
-	}
-
-	sprintf(ifr.lifr_name, "tap%d", unit_nr);
-	ifr.lifr_ppa = unit_nr;
-	ifr.lifr_flags = flags;
-
-	// We need to do this, but it is allowed to fail.
-	// No I don't understand it.
-	(void)ioctl(tap2_fd, SIOCSLIFNAME, &ifr);
-
-	ip_fd = open(devname, O_RDWR);
-	if (ip_fd < 0) {
-		close(tap2_fd);
-		return -1;
-	}
-
-	mux_id = ioctl(ip_fd, I_LINK, tap2_fd);
-	if (mux_id < 0) {
-		close(tap2_fd);
-		close(ip_fd);
-		return -1;
-	}
-
-	close(tap2_fd);
-
-	return ip_fd;
-}
-
-int solaris_setup_tap(nozzle_t nozzle, char *devname, int namelen)
-{
-	int tap_fd = -1;
-	static char tap_name[80];
-	int unit_nr;
-
-	tap_fd = open("/dev/tap", O_RDWR);
-	if (tap_fd < 0) {
-		return -EIO;
-	}
-
-	unit_nr = ioctl(tap_fd, TUNNEWPPA, -1);
-	if (unit_nr < 0) {
-		close(tap_fd);
-		return -EIO;
-	}
-
-	if (ioctl(tap_fd, I_SRDOPT, RMSGD) < 0) {
-		close(tap_fd);
-		return -EIO;
-	}
-
-	if (strlen(devname) == 0) {
-		sprintf(tap_name, "tap%d", unit_nr);
-		strncpy(devname, tap_name, namelen);
-	} else {
-		if (sscanf(devname, "tap%d", &unit_nr) == 1) {
-			struct strioctl strioc_ppa;
-			int ppa = unit_nr;
-			int newppa;
-			memset(&strioc_ppa, 0, sizeof(strioc_ppa));
-
-			strioc_ppa.ic_cmd = TUNNEWPPA;
-			strioc_ppa.ic_timout = 0;
-			strioc_ppa.ic_len = sizeof(ppa);
-			strioc_ppa.ic_dp = (char *)&ppa;
-			if ((newppa = ioctl(tap_fd, I_STR, &strioc_ppa)) < 0) {
-				return -errno;
-			}
-		} else {
-			return -EIO;
-		}
-	}
-
-	nozzle->ip_fd = link_proto(nozzle, unit_nr, "/dev/udp", IFF_IPV4);
-	if (nozzle->ip_fd < 0) {
-		close(tap_fd);
-		return -EIO;
-	}
-
-	nozzle->ip6_fd = link_proto(nozzle, unit_nr, "/dev/udp6", IFF_IPV6);
-	if (nozzle->ip6_fd < 0) {
-		close(tap_fd);
-		close(nozzle->ip_fd);
-		nozzle->ip_fd = -1;
-		return -EIO;
-	}
-
-	return tap_fd;
-}
-
-#endif
 
 /*
  * Exported public API
@@ -921,15 +123,6 @@ nozzle_t nozzle_open(char *devname, size_t devname_size, const char *updownpath)
 	int savederrno = 0;
 	nozzle_t nozzle = NULL;
 	char *temp_mac = NULL;
-#ifdef KNET_LINUX
-	struct ifreq ifr;
-#endif
-#ifdef KNET_BSD
-	uint16_t i;
-	long int nozzlenum = 0;
-	char curnozzle[IFNAMSIZ];
-	struct ifreq ifr;
-#endif
 
 	if (devname == NULL) {
 		errno = EINVAL;
@@ -946,30 +139,6 @@ nozzle_t nozzle_open(char *devname, size_t devname_size, const char *updownpath)
 		errno = E2BIG;
 		return NULL;
 	}
-
-#ifdef KNET_BSD
-	/*
-	 * BSD does not support named devices like Linux
-	 * but it is possible to force a nozzleX device number
-	 * where X is 0 to 255.
-	 */
-	if (strlen(devname)) {
-		if (strncmp(devname, "tap", 3)) {
-			errno = EINVAL;
-			return NULL;
-		}
-		errno = 0;
-		nozzlenum = strtol(devname+3, NULL, 10);
-		if (errno) {
-			errno = EINVAL;
-			return NULL;
-		}
-		if ((nozzlenum < 0) || (nozzlenum > 255)) {
-			errno = EINVAL;
-			return NULL;
-		}
-	}
-#endif
 
 	if (updownpath) {
 		/* only absolute paths */
@@ -991,54 +160,19 @@ nozzle_t nozzle_open(char *devname, size_t devname_size, const char *updownpath)
 
 	if (!lib_init) {
 		lib_cfg.head = NULL;
-#ifdef KNET_LINUX
-		int nlerr;
 
-		lib_cfg.nlsock = nl_socket_alloc();
-		if (!lib_cfg.nlsock) {
+		if (_platform_init(&lib_cfg) < 0) {
 			savederrno = errno;
 			goto out_error;
 		}
-		nlerr = nl_connect(lib_cfg.nlsock, NETLINK_ROUTE);
-		if (nlerr < 0) {
-			savederrno = nlerr_to_errno(nlerr);
-			goto out_error;
-		}
-		lib_cfg.ioctlfd = socket(AF_INET, SOCK_STREAM, 0);
-#endif
-#ifdef KNET_BSD
-		lib_cfg.ioctlfd = socket(AF_LOCAL, SOCK_DGRAM, 0);
+
+		lib_cfg.ioctlfd = socket(NOZZLE_SOCKET_DOMAIN, SOCK_DGRAM, 0);
 		if (lib_cfg.ioctlfd < 0) {
 			savederrno = errno;
+			_platform_fini(&lib_cfg);
 			goto out_error;
 		}
-		lib_cfg.ip_fd = socket(AF_INET, SOCK_DGRAM, 0);
-		if (lib_cfg.ip_fd < 0) {
-			savederrno = errno;
-			close(lib_cfg.ioctlfd);
-			goto out_error;
-		}
-		lib_cfg.ip6_fd = socket(AF_INET6, SOCK_DGRAM, 0);
-		if (lib_cfg.ip6_fd < 0) {
-			savederrno = errno;
-			close(lib_cfg.ioctlfd);
-			close(lib_cfg.ip_fd);
-			goto out_error;
-		}
-#endif
-#ifdef KNET_SOLARIS
-		lib_cfg.ioctlfd = socket(AF_INET, SOCK_DGRAM, IPPROTO_IP);
-		if (lib_cfg.ioctlfd < 0) {
-			savederrno = errno;
-			goto out_error;
-		}
-#endif
-#ifndef KNET_BSD
-		if (lib_cfg.ioctlfd < 0) {
-			savederrno = errno;
-			goto out_error;
-		}
-#endif
+
 		lib_init = 1;
 	}
 
@@ -1050,86 +184,10 @@ nozzle_t nozzle_open(char *devname, size_t devname_size, const char *updownpath)
 
 	memset(nozzle, 0, sizeof(struct nozzle_iface));
 
-#ifdef KNET_BSD
-	if (!strlen(devname)) {
-		for (i = 0; i < 256; i++) {
-			memset(&ifr, 0, sizeof(ifr));
-
-			snprintf(curnozzle, sizeof(curnozzle) - 1, "tap%u", i);
-			memmove(ifr.ifr_name, curnozzle, IFNAMSIZ);
-			if (ioctl(lib_cfg.ioctlfd, SIOCIFCREATE2, &ifr) < 0) {
-			        continue;
-			}
-			snprintf(curnozzle, sizeof(curnozzle) - 1, "/dev/tap%u", i);
-			nozzle->fd = open(curnozzle, O_RDWR);
-			savederrno = errno;
-			if (nozzle->fd > 0) {
-				break;
-			}
-			/* For some reason we can't open that device, keep trying
-			   but don't leave debris */
-			(void)ioctl(lib_cfg.ioctlfd, SIOCIFDESTROY, &ifr);
-			(void)ioctl(lib_cfg.ioctlfd, SIOCGIFFLAGS, &ifr);
-		}
-		snprintf(curnozzle, sizeof(curnozzle) -1 , "tap%u", i);
-	} else {
-		memmove(ifr.ifr_name, devname, IFNAMSIZ);
-		if (ioctl(lib_cfg.ioctlfd, SIOCIFCREATE2, &ifr) < 0) {
-			goto out_error;
-		}
-		snprintf(curnozzle, sizeof(curnozzle) - 1, "/dev/%s", devname);
-		nozzle->fd = open(curnozzle, O_RDWR);
-		savederrno = errno;
-		snprintf(curnozzle, sizeof(curnozzle) - 1, "%s", devname);
-	}
-	if (nozzle->fd < 0) {
-		savederrno = EBUSY;
-		goto out_error;
-	}
-	memmove(devname, curnozzle, IFNAMSIZ);
-	memmove(nozzle->name, curnozzle, IFNAMSIZ);
-#endif
-
-#ifdef KNET_SOLARIS
-	nozzle->fd = solaris_setup_tap(nozzle, devname, devname_size);
-	memmove(nozzle->name, devname, IFNAMSIZ);
-#endif
-
-#ifdef KNET_LINUX
-	if ((nozzle->fd = open("/dev/net/tun", O_RDWR)) < 0) {
+	if (_platform_create_tap(nozzle, devname, devname_size) < 0) {
 		savederrno = errno;
 		goto out_error;
 	}
-
-	memset(&ifr, 0, sizeof(struct ifreq));
-	memmove(ifname, devname, IFNAMSIZ);
-	ifflags = IFF_TAP | IFF_NO_PI;
-
-	/*
-	 * Use IFF_TUN_EXCL to prevent race conditions when creating named devices.
-	 * Without this flag, another process could create the same device name
-	 * between our check and creation, leading to unexpected behavior.
-	 * Available since Linux 3.4. Fallback to non-exclusive if not supported.
-	 */
-	if (strlen(devname) > 0) {
-#ifdef IFF_TUN_EXCL
-		ifflags |= IFF_TUN_EXCL;
-#endif
-	}
-
-	if (ioctl(nozzle->fd, TUNSETIFF, &ifr) < 0) {
-		savederrno = errno;
-		goto out_error;
-	}
-
-	if ((strlen(devname) > 0) && (strcmp(devname, ifname) != 0)) {
-		savederrno = EBUSY;
-		goto out_error;
-	}
-
-	memmove(devname, ifname, IFNAMSIZ);
-	memmove(nozzle->name, ifname, IFNAMSIZ);
-#endif
 	nozzle->default_mtu = get_iface_mtu(nozzle);
 	if (nozzle->default_mtu < 0) {
 		savederrno = errno;
@@ -1206,10 +264,8 @@ int nozzle_close(nozzle_t nozzle)
 		free(ip);
 		ip = ip_next;
 	}
-#ifdef KNET_SOLARIS
-	close(nozzle->ip_fd);
-	close(nozzle->ip6_fd);
-#endif
+
+	_platform_close_tap(nozzle);
 
 	destroy_iface(nozzle);
 
@@ -1298,11 +354,7 @@ out_clean:
 int nozzle_set_up(nozzle_t nozzle)
 {
 	int err = 0, savederrno = 0;
-#ifdef KNET_SOLARIS
-	struct lifreq ifr;
-#else
-	struct ifreq ifr;
-#endif
+	nozzle_ifreq ifr;
 
 	savederrno = pthread_mutex_lock(&config_mutex);
 	if (savederrno) {
@@ -1347,11 +399,7 @@ out_clean:
 int nozzle_set_down(nozzle_t nozzle)
 {
 	int err = 0, savederrno = 0;
-#ifdef KNET_SOLARIS
-	struct lifreq ifr;
-#else
-	struct ifreq ifr;
-#endif
+	nozzle_ifreq ifr;
 
 	savederrno = pthread_mutex_lock(&config_mutex);
 	if (savederrno) {
@@ -1451,11 +499,6 @@ out_clean:
 int nozzle_set_mac(nozzle_t nozzle, const char *ether_addr)
 {
 	int err = 0, savederrno = 0;
-#ifdef KNET_SOLARIS
-	dlpi_handle_t dlpi_handle;
-#else
-	struct ifreq ifr;
-#endif
 
 	if (!ether_addr) {
 		errno = EINVAL;
@@ -1474,51 +517,9 @@ int nozzle_set_mac(nozzle_t nozzle, const char *ether_addr)
 		goto out_clean;
 	}
 
-#ifdef KNET_LINUX
-	memset(&ifr, 0, sizeof(struct ifreq));
-	memmove(ifname, nozzle->name, IFNAMSIZ);
-	err = ioctl(lib_cfg.ioctlfd, SIOCGIFHWADDR, &ifr);
-	if (err) {
-		savederrno = errno;
-		goto out_clean;
-	}
-
-	memmove(ifr.ifr_hwaddr.sa_data, ether_aton(ether_addr), ETH_ALEN);
-
-	err = ioctl(lib_cfg.ioctlfd, SIOCSIFHWADDR, &ifr);
+	err = _platform_set_mac(nozzle, ether_addr);
 	savederrno = errno;
-#endif
-#ifdef KNET_BSD
-	memset(&ifr, 0, sizeof(struct ifreq));
-	memmove(ifname, nozzle->name, IFNAMSIZ);
-	err = ioctl(lib_cfg.ioctlfd, SIOCGIFADDR, &ifr);
-	if (err) {
-		savederrno = errno;
-		goto out_clean;
-	}
 
-	memmove(ifr.ifr_addr.sa_data, ether_aton(ether_addr), ETHER_ADDR_LEN);
-	ifr.ifr_addr.sa_len = ETHER_ADDR_LEN;
-
-	err = ioctl(lib_cfg.ioctlfd, SIOCSIFLLADDR, &ifr);
-	savederrno = errno;
-#endif
-#ifdef KNET_SOLARIS
-	err = dlpi_open(nozzle->name, &dlpi_handle, 0);
-	if (err != DLPI_SUCCESS) {
-		err = -1;
-		goto out_clean;
-	}
-	err = dlpi_set_physaddr(dlpi_handle, DL_CURR_PHYS_ADDR,
-				ether_aton(ether_addr), ETHERADDRL);
-
-	if (err != DLPI_SUCCESS) {
-		err = -1;
-		dlpi_close(dlpi_handle);
-		goto out_clean;
-	}
-	dlpi_close(dlpi_handle);
-#endif
 out_clean:
 	pthread_mutex_unlock(&config_mutex);
 	errno = savederrno;
@@ -1614,11 +615,7 @@ int nozzle_set_mtu(nozzle_t nozzle, const int mtu)
 {
 	int err = 0, savederrno = 0;
 	struct nozzle_ip *tmp_ip;
-#ifdef KNET_SOLARIS
-	struct lifreq ifr;
-#else
-	struct ifreq ifr;
-#endif
+	nozzle_ifreq ifr;
 
 	if (!mtu) {
 		errno = EINVAL;
@@ -1647,25 +644,17 @@ int nozzle_set_mtu(nozzle_t nozzle, const int mtu)
 	memmove(ifname, nozzle->name, IFNAMSIZ);
 	ifmtu = mtu;
 
-#ifdef KNET_SOLARIS
-	err = ioctl(nozzle->ip_fd, SIOCSLIFMTU, &ifr);
-#else
-	err = ioctl(lib_cfg.ioctlfd, SIOCSIFMTU, &ifr);
-#endif
+	err = ioctl(NOZZLE_IOCTL_FD, NOZZLE_SET_MTU, &ifr);
 	if (err) {
 		savederrno = errno;
 		goto out_clean;
 	}
 
 	if ((nozzle->current_mtu < 1280) && (mtu >= 1280)) {
-#ifdef KNET_SOLARIS
-		int secondary = 1;
-#else
-		int secondary = 0;
-#endif
 		tmp_ip = nozzle->ip;
 		while(tmp_ip) {
 			if (tmp_ip->domain == AF_INET6) {
+				int secondary = NOZZLE_IPV6_IS_SECONDARY(tmp_ip->domain);
 				err = _set_ip(nozzle, IP_ADD, tmp_ip->ipaddr, tmp_ip->prefix, secondary);
 				if (err) {
 					savederrno = errno;
@@ -1743,11 +732,7 @@ int nozzle_add_ip(nozzle_t nozzle, const char *ipaddr, const char *prefix)
 		/* We make all Solaris IP6 addresses secondary's (using addif)
 		 * otherwise we can't remove the last one
 		 */
-		if (nozzle->ip
-#ifdef KNET_SOLARIS
-		   || ip->domain == AF_INET6
-#endif
-			) {
+		if (nozzle->ip || NOZZLE_IPV6_IS_SECONDARY(ip->domain)) {
 			secondary = 1;
 		}
 		err = _set_ip(nozzle, IP_ADD, ipaddr, prefix, secondary);
@@ -1811,13 +796,8 @@ int nozzle_del_ip(nozzle_t nozzle, const char *ipaddr, const char *prefix)
 	if ((ip->domain == AF_INET6) && (get_iface_mtu(nozzle) < 1280)) {
 		err = 0;
 	} else {
-		int secondary = 0;
-#ifdef KNET_SOLARIS
 		/* On Solaris, all IPv6 addresses are secondary logical interfaces */
-		if (ip->domain == AF_INET6) {
-			secondary = 1;
-		}
-#endif
+		int secondary = NOZZLE_IPV6_IS_SECONDARY(ip->domain);
 		err = _set_ip(nozzle, IP_DEL, ipaddr, prefix, secondary);
 		savederrno = errno;
 	}
